Complete §2a model precedence: requested_model + stop_reason + usage_kind
Decompose + extend:
- Extract the audit-write functions (agent_log_call, agent_record_token_audit,
  agent_set_ingress_source) from agent_runtime.c (at the 2000-line limit) into a
  new agent_logging.c. Wired into AGENT_SRCS and the test object set.
- token_audit gains requested_model, stop_reason, and usage_kind columns (additive
  migration via db1_reconcile_columns; usage_kind defaults to "realized").
- agent_result_t carries requested_model + stop_reason; parsed_response_t carries
  stop_reason. The OpenAI/Anthropic parsers capture finish_reason/stop_reason, and
  the server wrappers propagate it to the result.
- The ingress handlers record the client's requested model separately from the
  served model (e.g. Claude Code asks for a model, aimee serves its primary), and
  the captured stop_reason, via the audit helper.

This completes review #4's billable-model precedence (provider-reported > served,
requested recorded separately) for the ingress and server agent paths. usage_kind
is in place for the realized/estimated/avoided distinction (readers gain the
filter when dedup/estimate rows arrive).

Test: ingress row records requested_model + stop_reason + usage_kind=realized.
Server build + full unit suite + lint + schema-sync green.
