+++ src/Makefile
@@ -326,7 +326,7 @@ DATA_SRCS = memory_core.c memory_logic.c memory_effective.c memory_health.c memo
 DATA_OBJS = $(DATA_SRCS:%.c=$(OBJDIR)/%.o)
 
 # --- Layer 2: Agent execution (agent loop, tools, HTTP, policy) ---
-AGENT_SRCS = server/agent_runtime.c server/agent_parallel.c server/agent_runtime_liveness.c server/agent_fallback.c server/agent_context_budget.c server/agent_bridge.c server/tool_call_args.c server/agent_request_shaping.c server/agent_adapter.c server/primary_session_adapter.c server/agent_policy.c server/agent_policy_intercept.c server/agent_eval.c server/agent_eval_memory_support.c server/agent_eval_baseline.c server/agent_eval_benchmarks.c server/agent_coord.c \
+AGENT_SRCS = server/agent_runtime.c server/agent_logging.c server/agent_parallel.c server/agent_runtime_liveness.c server/agent_fallback.c server/agent_context_budget.c server/agent_bridge.c server/tool_call_args.c server/agent_request_shaping.c server/agent_adapter.c server/primary_session_adapter.c server/agent_policy.c server/agent_policy_intercept.c server/agent_eval.c server/agent_eval_memory_support.c server/agent_eval_baseline.c server/agent_eval_benchmarks.c server/agent_coord.c \
              server/agent_tasks.c server/agent_tools.c server/computer_use.c server/script_runner.c server/script_rpc.c toolset.c server/tool_args_coerce.c server/tool_schema_sanitizer.c server/osv_check.c server/http_retry.c server/failover.c \
              server/web_search.c server/token_tracker.c server/token_tracker_registry.c server/middleware.c server/agent_pipeline.c \
              server/process_mgr.c server/agent_loop.c server/liveness.c server/cli_session.c server/cli_codex.c server/provider_cli_adapter.c server/cli_claude.c server/cli_gemini.c server/cli_mistral.c server/cli_acp.c server/session_compact.c server/compact_prune.c server/context_engine.c \
